Internal: Fix build warnings 2/2

+++ src/game/PlayerBots/playerbot/strategy/actions/ChooseTravelTargetAction.cpp
@@ -1393,7 +1393,7 @@ bool RequestQuestTravelTargetAction::Execute(Event& event)
             if (!flag)
                 continue;
 
-            destinationFetches.push_back({ flag, questId, 1000 + (bot->GetLevel() * bot->GetLevel()) * 75 });
+            destinationFetches.push_back({ flag, static_cast<int32>(questId), static_cast<float>(1000 + (bot->GetLevel() * bot->GetLevel()) * 75) });
 
             if (onlyClassQuest && destinationFetches.size() > 1) //Only do class quests if we have any.
             {
